Cellular feature, ESP-IDF only: PPP-level integration.
This commit introduces the concept of a PPP interface to ubxlib, currently only for the ESP-IDF platform and with a cellular transport.  This allows the native application protocols of the platform to be used with cellular.

The PPP interface is defined as a new port API, which includes a default implementation that does nothing so that no existing ports are adversely affected by it.  This port API is, internally, called from the cellular code, such that when the cellular network is up the PPP interface of the port layer is automatically activated, connecting into the PPP layer at the bottom of the IP stack of the platform (LWIP in the ESP-IDF case).

This is currently only supported with SARA-R5 and SARA-R422 cellular modules.  U_CFG_PPP_ENABLE must be define when building ubxlib to enable PPP.

Of course, for this to integrate with the platform there will be additional build requirements related to the platform itself (enabling the right components of the platform etc.): see the README.md in the ESP-IDF platform directory for details.

An example is added to the sockets examples showing how to use native ESP-IDF sockets with PPP.